refactor: relocate KoColor domain models to core ritual package
This commit moves several model classes from the KoColor application module to a centralized `core:model` module. This promotes code reuse and standardizes the ritual-related data structures within the core library.

- **Model Relocation**:
    - Moved `BeautyRoutine`, `ClothingItem`, `CosmeticItem`, `GlassButtonState`, and `HarmonizedLook` from `applications:kocolor:model` to `com.zoewave.probase.core.model.ritual`.
    - Renamed `FashionProfile.kt` to `FashionModels.kt` during the relocation to the core module.
- **Package Updates**:
    - Updated the package declarations for all moved files to `com.zoewave.probase.core.model.ritual`.
